handle audio/video decoder error
When there is input data and no output data to the end of the stream, it will
send GST_ELEMENT_ERROR, So the clips playing will quit.
However, if only one of the tracks is corrupt, there is no need to quit other
tracks playing.

The patch comments the GST_ELEMENT_ERROR() and just add GST_ERROR_OBJECT()
information instead.

https://bugzilla.gnome.org/show_bug.cgi?id=741542

Upstream Status: Pending

Signed-off-by: Lyon Wang <lyon.wang@freescale.com>