allocators: Add GstDmabufMeta
This meta can be added to the allocation query to indicate support
for dmabuf memory.  Optionally the drm_modifier can also be set in
the allocation query to indicate that the element supports some
certain set of drm modifiers (swizzling/tiling/compression/etc).
When the meta is attached to a buffer, the drm_modifier should
only be one which the src and sink both indicated support for in
the allocation query.

https://bugzilla.gnome.org/show_bug.cgi?id=779146
(cherry picked from commit 896fbcf7a56ba8eb8ac79f6526405aa1d3561bb3)